pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "strategylab"
version = "0.1.0"
description = "Extensible trading strategy backtesting and monitoring platform"
requires-python = ">=3.11"
dependencies = [
    "yfinance>=0.2.31",
    "pandas>=2.0",
    "numpy>=1.24",
    "scipy>=1.11",
    "matplotlib>=3.7",
    "jinja2>=3.1",
    "click>=8.1",
    "rich>=13.0",
    "schedule>=1.2",
    "pytz>=2023.3",
    "pyyaml>=6.0",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
]

[project.scripts]
strategylab = "strategylab.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
